Move old team builder view to a component
